adding delete work endpoint

+++ src/main/java/br/com/taskmate/controller/UserController.java
@@ -1,17 +1,25 @@
 package br.com.taskmate.controller;
 
+import br.com.taskmate.infra.security.TokenService;
 import br.com.taskmate.model.user.Client;
 import br.com.taskmate.model.user.User;
 import br.com.taskmate.model.user.Worker;
 import br.com.taskmate.service.UserService;
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+
 
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.tags.Tag;
+
+@Tag(name = "TaskMate API")
 @RestController
 @RequestMapping("/user")
 public class UserController {
@@ -19,6 +27,10 @@ public class UserController {
     @Autowired
     private UserService userService;
 
+    @Autowired
+    private TokenService tokenService;
+
+    @Operation(description = "Get all users")
     @GetMapping("/all")
     public List<User> getAllUsers() {
         return userService.findAllUsers();
@@ -52,8 +64,15 @@ public class UserController {
         }
     }
 
-    @DeleteMapping("/delete/{id}")
-    public ResponseEntity<Void> deleteUser(@PathVariable UUID id) {
+    @DeleteMapping("/deleteAccount/{id}")
+    public ResponseEntity<Void> deleteUser(@PathVariable UUID id, HttpServletRequest request) {
+        String token = request.getHeader("Authorization").replace("Bearer ", "");
+        String username = tokenService.validateToken(token);
+
+        if (username.isEmpty()) {
+            return ResponseEntity.status(401).build();
+        }
+
         if (userService.findUserById(id).isPresent()) {
             userService.deleteUser(id);
             return ResponseEntity.noContent().build();
